refactor: move to test and util file
